feat: create class for timer mode and add configurable time to the counter
