+++ index.html
@@ -55,6 +55,6 @@
       </div>
     </main>
 
-    <script src="js/Counter.js"></script>
+    <script src="js/index.js"></script>
   </body>
 </html>
